Rename all child-parent feature files and classes to one naming model
