+++ src/presentation/http/http-api.ts
@@ -190,7 +190,7 @@ export default class HttpApi implements Api {
       prefix: '/note',
       noteService: domainServices.noteService,
       noteSettingsService: domainServices.noteSettingsService,
-      noteRelationshipService: domainServices.noteRelationshipService,
+      noteRelationsService: domainServices.noteRelationshipService,
     });
 
     await this.server?.register(NoteListRouter, {
