+++ src/repository/storage/noteRelations.storage.ts
@@ -1,0 +1,6 @@
+import NoteRelationsSequelizeStorage from './postgres/orm/sequelize/noteRelations.js';
+
+/**
+ * Current note relationship storage
+ */
+export default NoteRelationsSequelizeStorage;
