--- src/repository/storage/noteRelationshp.storage.ts
@@ -1,6 +1,0 @@
-import NoteRelationshipSequelizeStorage from './postgres/orm/sequelize/noteRelationship.js';
-
-/**
- * Current note relationship storage
- */
-export default NoteRelationshipSequelizeStorage;
